selinux: fix check for transient units
SELinux does not have a path to check for a snapshot service creation.
This ends up giving us a bogus check.

On snapshot creation we should check if the remote process type, has the
ability to start a service with the type that systemd is running with.

Based on patch from Vaclav Pavlin and Dan Walsh
http://lists.freedesktop.org/archives/systemd-devel/2013-November/014021.html

RHEL only
Resolves: #1255129