manager: reduce complexity of unit_gc_sweep
When unit is marked as UNSURE, we are trying to find if it state was
changed over and over again. But only thing that could change is that
some of its dependencies was add to cleanup queue, so it might go to BAD
state, but that does not matter since we remove both UNSURE and BAD
units.

This is a problem on machines with a lot of initscripts with different
starting priority, since those units will reference each other and the
original algorithm might get to n! complexity.

rhel-only
not in upstream yet
Resolves: #1344556